@@ -2,15 +2,14 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace BitcoinBook.Test
-{
-    public abstract class FakeHttpMessageHandler : HttpMessageHandler
-    {
-        public abstract HttpResponseMessage Send(HttpRequestMessage request);
+namespace BitcoinBook.Test;
 
-        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(Send(request));
-        }
+public abstract class FakeHttpMessageHandler : HttpMessageHandler
+{
+    public abstract HttpResponseMessage Send(HttpRequestMessage request);
+
+    protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+    {
+        return Task.FromResult(Send(request));
     }
 }
