@@ -2,28 +2,27 @@
 using System.Net.Http;
 using Xunit;
 
-namespace BitcoinBook.Test
+namespace BitcoinBook.Test;
+
+public class SignerMapTests
 {
-    public class SignerMapTests
+    readonly SignerMap signerMap;
+
+    public SignerMapTests()
     {
-        readonly SignerMap signerMap;
+        var fetcher = new TransactionFetcher(new HttpClient());
+        var hasher = new TransactionHasher(null!);
 
-        public SignerMapTests()
-        {
-            var fetcher = new TransactionFetcher(new HttpClient());
-            var hasher = new TransactionHasher(null!);
+        signerMap = new SignerMap(
+            new PayToPubKeySigner(fetcher, hasher), 
+            new PayToPubKeyHashSigner(fetcher, hasher));
+    }
 
-            signerMap = new SignerMap(
-                new PayToPubKeySigner(fetcher, hasher), 
-                new PayToPubKeyHashSigner(fetcher, hasher));
-        }
-
-        [Theory]
-        [InlineData(typeof(PayToPubKeySigner), ScriptType.PayToPubKey)]
-        [InlineData(typeof(PayToPubKeyHashSigner), ScriptType.PayToPubKeyHash)]
-        public void SignerTest(Type signerType, ScriptType scriptType)
-        {
-            Assert.IsType(signerType, signerMap[scriptType]);
-        }
+    [Theory]
+    [InlineData(typeof(PayToPubKeySigner), ScriptType.PayToPubKey)]
+    [InlineData(typeof(PayToPubKeyHashSigner), ScriptType.PayToPubKeyHash)]
+    public void SignerTest(Type signerType, ScriptType scriptType)
+    {
+        Assert.IsType(signerType, signerMap[scriptType]);
     }
 }
